fixed some ts stuff
